requirements.txt
bs4
requests_tor==1.4
